+++ src/katalyst_core/utils/syntax_checker.py
@@ -1,7 +1,7 @@
 import tempfile
 import os
 from tree_sitter_languages import get_parser
-from coding_agent.config import EXT_TO_LANG
+from app.config import EXT_TO_LANG
 
 # --- Syntax Checking Utilities ---
 # This module provides syntax checking for multiple languages.
